+++ Groceries/AppIntents/OpenGroceriesIntent.swift
@@ -13,32 +13,3 @@ struct OpenGroceriesIntent: AppIntent {
         return .result()
     }
 }
-
-import Foundation
-
-// MARK: - App Shortcuts
-
-struct OpenGroceriesShortcut: AppShortcutsProvider {
-    
-    @AppShortcutsBuilder
-    static var appShortcuts: [AppShortcut] {
-        AppShortcut(
-            intent: OpenGroceriesIntent(),
-            phrases: [
-                "Open \(.applicationName)",
-                "See \(.applicationName)",
-                "See my \(.applicationName)",
-                "What is my \(.applicationName)?",
-                "What's my \(.applicationName)?",
-                "What is on my \(.applicationName)?",
-                "What's on my \(.applicationName)?",
-                "Show \(.applicationName)",
-                "Show my \(.applicationName)",
-                "Show me my \(.applicationName)",
-            ],
-            shortTitle: LocalizedStringResource(stringLiteral: "Open groceries"),
-            systemImageName: "apple.logo"
-        )
-    }
-    
-}
